Add files via upload
I see no changes, wake up in the morning and I ask myself
Is life worth livin'? Should I blast myself?
I'm tired of bein' poor and, even worse, I'm black
My stomach hurts so I'm lookin' for a purse to snatch
Cops give a damn about a negro
Pull the trigger, kill a nigga, he's a hero
"Give the crack to the kids, who the hell cares?
One less hungry mouth on the welfare"
First ship 'em dope and let 'em deal to brothers
Give 'em guns, step back, watch 'em kill each other
"It's time to fight back," that's what Huey said
Two shots in the dark, now Huey's dead
I got love for my brother
But we can never go nowhere unless we share with each other
We gotta start makin' changes
Learn to see me as a brother instead of two distant strangers
And that's how it's supposed to be
How can the Devil take a brother if he's close to me? Uh
I'd love to go back to when we played as kids
But things change, and that's the way it is
See upcoming rap shows
Get tickets for your favorite artists

(Come on, come on)
That's just the way it is (Changes)
Things'll never be the same
That's just the way it is (That's the way it is, what?)
Aww, yeah-yeah (Hear me)
(Oh my, oh my, come on, come on)
That's just the way it is (That's just the way it is, the way it is)
Things'll never be the same
(Never be the same, yeah, yeah, yeah, aww, yeah)
That's just the way it is (Way it is)
Aww, yeah (Come on, come on)

I see no changes, all I see is racist faces
Misplaced hate makes disgrace to races
We under, I wonder what it takes to make this
One better place, let's erase the wasted
Take the evil out the people, they'll be actin' right
'Cause both black and white are smokin' crack tonight
And the only time we chill is when we kill each other (Kill each other)
It takes skill to be real, time to heal each other
And although it seems heaven-sent
We ain't ready to see a black president, uh (Oh-ooh)
It ain't a secret, don't conceal the fact
The penitentiary's packed and it's filled with blacks
But some things will never change (Never change)
Try to show another way, but you stayin' in the dope game (Ooh)
Now tell me, what's a mother to do?
Bein' real don't appeal to the brother in you (Yeah)
You gotta operate the easy way
"I made a G today," but you made it in a sleazy way
Sellin' crack to the kids (Oh-oh), "I gotta get paid" (Oh)
Well hey, well that's the way it is

(Come on, come on)
That's just the way it is (Changes)
Things'll never be the same
That's just the way it is (That's the way it is, what?)
Aww, yeah (Hear me)
(Oh my, oh my, come on, come on)
That's just the way it is (That's just the way it is, the way it is)
Things'll never be the same
(Never be the same, yeah, yeah, yeah, aww, yeah)
That's just the way it is (Way it is)
Aww, yeah (Aww, yeah, aww, yeah)

We gotta make a change
It's time for us as a people to start makin' some changes
Let's change the way we eat
Let's change the way we live
And let's change the way we treat each other
You see, the old way wasn't workin'
So it's on us to do what we gotta do to survive

And still I see no changes, can't a brother get a little peace?
It's war on the streets and the war in the Middle East (Ooh, yeah)
Instead of war on poverty
They got a war on drugs so the police can bother me
And I ain't never did a crime I ain't have to do
But now I'm back with the facts, givin' it back to you (Ooh)
Don't let 'em jack you up, back you up
Crack you up and pimp-smack you up
You gotta learn to hold your own
They get jealous when they see you with your mobile phone
But tell the cops they can't touch this
I don't trust this, when they try to rush, I bust this
That's the sound of my tool, you say it ain't cool
My mama didn't raise no fool (Oh)
And as long as I stay black, I gotta stay strapped
And I never get to lay back
'Cause I always got to worry 'bout the payback
Some buck that I roughed up way back
Comin' back after all these years
"Rat-a-tat-tat-tat-tat," that's the way it is
(Uh)
That's just the way it is (Just the way it is, yeah, yeah, yeah)
Things'll never be the same (Yeah)
That's just the way it is (The way it is)
Aww, yeah (Some things will never change, oh my)
(I'm tryna make a change)
(You're my brother, you're my sister, yeah)
That's just the way it is (The way it is, the way it is)
Things'll never be the same (You're my brother, you're my sister)
That's just the way it is, aww, yeah
Some things'll never change